Rewrite pick to nif and move to c_src
Makes it a more standard erlang application.

Also make make it thread safe.
