+++ .dir-locals.el
@@ -11,4 +11,4 @@
   (c-basic-offset . 2))
  (c-mode
   (indent-tabs-mode . nil)
-  (c-basic-offset . 4)))
+  (c-basic-offset . 2)))
