+++ .gitignore
@@ -31,6 +31,9 @@
 # /fonts/
 /fonts/*.wingsfont
 
+# /priv/
+/priv/*
+
 # /plugins/
 /plugins/*
 
@@ -39,10 +42,6 @@
 
 # /plugins_src/autouv/
 /plugins_src/autouv/autouv_en.lang
-
-# /plugins_src/accel/
-/plugins_src/accel/*.obj
-/plugins_src/accel/*.pdb
 
 # /plugins_src/fbx/
 /plugins_src/fbx/*.o
