+++ .travis.yml
@@ -33,7 +33,7 @@ script:
 
 after_success:
   - dialyzer --build_plt --apps erts kernel stdlib wx tools xmerl cl --statistics
-  - dialyzer -n -Wno_improper_lists ebin/*.beam plugins/accel/*.beam plugins/autouv/*.beam plugins/commands/*.beam plugins/default/*.beam plugins/import_export/*.beam plugins/primitives/*.beam --statistics
+  - dialyzer -n -Wno_improper_lists ebin/*.beam plugins/autouv/*.beam plugins/commands/*.beam plugins/default/*.beam plugins/import_export/*.beam plugins/primitives/*.beam --statistics
 
 # Needs a compiled cl-1.2.3
 after_script:
